+++ lib/save_it/video_upload.ex
@@ -3,6 +3,7 @@ defmodule SaveIt.VideoUpload do
 
   require Logger
 
+  alias SaveIt.DownloadedFileName
   alias SaveIt.VideoMetadata
 
   @telegram_cover_max_dimension 320
@@ -96,12 +97,7 @@ defmodule SaveIt.VideoUpload do
 
   defp cover_dimensions(_metadata), do: {:error, :missing_video_display_dimensions}
 
-  defp cover_file_name(file_name) do
-    file_name
-    |> Path.basename()
-    |> Path.rootname()
-    |> Kernel.<>("-cover.jpg")
-  end
+  defp cover_file_name(_file_name), do: DownloadedFileName.random("cover.jpg")
 
   defmodule FFmpegFaststart do
     @moduledoc false